requirements/testing.txt
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-isort==0.3.1
pytest-sugar==0.9.2
pytest==4.3.0
